hps/exploits/ECSC_valve_firmware_gen/requirements.txt
tlv8==0.10.0
pycryptodome==3.15.0